edited_prana/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
django-filter==23.3
django-unfold==0.64.2
django-cors-headers==4.0.0
Pillow==10.0.0